+++ sdk/core/src/services/BloomVaultStorageService.ts
@@ -6,7 +6,7 @@ import { toRpcSig, ecsign } from 'ethereumjs-util'
 import { IPlatformEncryptionTools } from '../shared/interfaces'
 import SdkErrorFromCode from '../shared/SdkErrorFromCode'
 import { FetchCredentialsPaginationOptions } from '../dto/shared.dto'
-import { createApiServiceHeaders, isW3cCredential } from '../_helpers'
+import { extractSDKVersion, isW3cCredential } from '../_helpers'
 
 const keccak256 = require('keccak256')
 const secp256k1 = require('secp256k1')
@@ -58,7 +58,7 @@ export default class BloomVaultStorageService {
     this._vaultApiService = new BloomVaultApiService({
       vaultUrl: options.vaultUrl,
       accessApiKey: options.accessApiKey,
-      headers: createApiServiceHeaders(),
+      sdkVersion: extractSDKVersion(),
     })
   }