+++ common-libs/internal-api-clients/CHANGELOG.md
@@ -1,5 +1,7 @@
 # release 1.0.0-beta.6 (2021-07-26)
-  * Added headers to GenericApiService constructor options
+  * Moved headers management to `helpers` module
+  * Added `X-SDK-Version` header
+  * Added optional `sdkVersion` parameter to GenericApiService constructor options
 # release 1.0.0-beta.5 (2021-07-23)
   * Fixed passing empty `types` to `searchCredentials` in `AffinidiVaultApiService`
 # release 1.0.0-beta.4 (2021-07-23)